backend/LIVIS/requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.10
billiard==3.3.0.23
blis==0.4.1
catalogue==1.0.0
celery==3.1.26.post2
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
cymem==2.0.3
Cython==0.29.21
Django==3.0.8
django-celery==3.3.1
django-cors-headers==3.4.0
djangorestframework==3.11.0
dnspython==1.16.0
drf-yasg==1.17.1
eventlet==0.27.0
fastai==2.0.10
fastcore==1.0.9
fastprogress==1.0.0
future==0.18.2
gevent==1.5.0
greenlet==0.4.16
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
kombu==3.0.37
MarkupSafe==1.1.1
matplotlib==3.3.0
monotonic==1.5
murmurhash==1.0.2
numpy==1.19.1
opencv-python==4.3.0.36
packaging==20.4
pandas==1.1.0
Pillow==7.2.0
plac==1.1.3
preshed==3.0.2
pymongo==3.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
redis==2.10.6
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
spacy==2.3.2
sqlparse==0.3.1
srsly==1.0.2
thinc==7.4.1
threadpoolctl==2.1.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.48.2
uritemplate==3.0.1
urllib3==1.25.10
wasabi==0.8.0
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
